+++ NEWS.md
@@ -1,6 +1,7 @@
 # dbx 0.2.1 [unreleased]
 
 - Added `sslmode` and `sslrootcert` for RPostgreSQL
+- Added `pool` option to `dbxConnect`
 
 # dbx 0.2.0